refactor(e2e): merge setup and auth into single serial test flow (#94)
Combine setup and login tests into one describe.serial block so they
run deterministically: setup runs first (skips if already done), then
login tests always execute using a freshly registered test user.
